Phase 1c: double LBPG to 1024 on x86_64; widen free-list links to long
The free-list link encoding in the allocator and GC stored pointer
values into 4-byte int slots:
   *(int *)slot = (int)next_free;
This worked when sizeof(int) == sizeof(void *) (32-bit) and broke
silently on x86_64. The fix has two parts.

(1) LBPG doubled to 1024 on linux_x86_64 (option B from PortPlan).
This keeps every *SPP and type_len constant valid -- on 32-bit each
"slot" was N ints * 4 bytes; on 64-bit it becomes N longs * 8 bytes,
and the page doubled to match. Without this change, HUNK128 (128
slots) couldn't fit in a 512-byte page at all on 64-bit.

  * global.h: #if linux_x86_64 LBPG=1024 #else 512.
  * global.h: ATOX shifts by 10 on x86_64 (was 9).
  * config.h: TTSIZE halved to 32768 to keep the heap at 32 MB.
  * alloc.c: BITLONGS unified to TTSIZE*4 for both platforms -- the
    bitmap density (1 int per 256 bytes) is per-long, not per-page.

(2) Free-list links widened from int to long throughout alloc.c.

  * get_more_space: int *loop, *temp -> long *loop, *temp; (int)
    casts on pointer values -> (long).
  * gc1 (GC sweep): same widening for `point`, `start`, `head`,
    `lim`, `tempp`. The constants ("4 bitmap-ints per page",
    "32 longs = 16 dtprs", "advance start by 128 per page") happen
    to be correct on both platforms.
  * pruneb: instead of writing into temp->s.I (the 4-byte int field
    of struct sdot), snap temp->s.CDR first and write the link into
    the first long of the slot, matching get_more_space's idiom.
  * Globals widened: bind_lists (compiler literal-table chain),
    beginsweep (start of sweepable heap). Cascaded through extern
    declarations in lam8.c, lamgc.c, sysat.c.

alloc.c is now warning-free. Total kernel warnings: 23 (down from
38). All 37 .c files still compile to x86_64 ELF. Remaining warnings
all in eval2.c (Phase 1g callg ABI), divbig.c/pbignum.c (Phase 1h
bignum redesign), and one pragma-resistant cast in lam8.c.

+++ franz/data.c
@@ -152,7 +152,7 @@ int trleft = 0;			/* number of entries left in current table */
 
 /* globals from sysat.c  */
 
-int *beginsweep;		/* place for sweeper to begin 		*/
+long *beginsweep;		/* place for sweeper to begin (widened from int* for 64-bit) */
 int initflag = TRUE;		/* inhibit gcing initially		*/
 int tgcthresh = 15;
 int page_limit = (5 * TTSIZE) / 6;
@@ -169,7 +169,12 @@ lispval lastrtab;
 char purepage[TTSIZE];
 int fakettsize = TTSIZE - 8;
 int gcstrings;				/*  Do we mark and sweep strings? */
-int  *bind_lists = (int *) CNIL;	/*  lisp data for compiled code */
+/* bind_lists is a list of compiler-linked literal tables. Each entry
+ * holds a Lisp value (a pointer); on 64-bit Linux those don't fit in
+ * `int`, so the type widened from int* to long*. The list is empty
+ * unless code is loaded via fasl, which we don't build on x86_64.
+ */
+long *bind_lists = (long *) CNIL;
 
 
 struct str_x str_current[2];		/*  next free string spaces */
